Move Abstract to its own module
